refactor: dry part 2
