+++ docs/.storybook/decorators/neighbouring-stacking-context/neighbouring-stacking-context.js
@@ -1,5 +1,6 @@
 import Card from '@commercetools-uikit/card';
 import styled from '@emotion/styled';
+import { boolean } from '@storybook/addon-knobs/react';
 
 const StackingContext = styled.div`
   width: 300px;
@@ -8,12 +9,20 @@ const StackingContext = styled.div`
   z-index: 2;
 `;
 
-const NeighbouringStackingContext = () => (
-  <StackingContext>
-    <Card theme="dark">
-      Stacking context with <code>z-index: 2</code>
-    </Card>
-  </StackingContext>
-);
+const NeighbouringStackingContext = () => {
+  const isActive = boolean(
+    'menuPortalZIndex-show-neighbouring-stacking-context',
+    false
+  );
+  return (
+    isActive && (
+      <StackingContext>
+        <Card theme="dark">
+          Stacking context with <code>z-index: 2</code>
+        </Card>
+      </StackingContext>
+    )
+  );
+};
 
 export default NeighbouringStackingContext;
