--- docs/.storybook/utils/get-menu-portal-target-value.js
@@ -1,6 +1,0 @@
-export const getMenuPortalTargetValue = (menuPortalTarget) => {
-  if (menuPortalTarget === 'document.body') {
-    return document.body;
-  }
-  return undefined;
-};
